fix: replace `\r\n` in blogposts with `\n` after merge
This is done to consistently enforce the Unix `\n` (linefeed)
line ending convention in the repo even if blogposts where merged
with Windows `\r\n` line endings. It also allows the script to
easily appropriately handle such files, as before it was unable
to properly parse it because it assumed files always were `\n`.